scraper/requirements.txt
python-jobspy
psycopg2-binary
sqlalchemy
schedule
python-dotenv
